requirements.txt
# ───────────────────────────────────────────────────────
# Agri AI Backend — Production Dependencies
# Install with: pip install --no-cache-dir -r requirements.txt
# ───────────────────────────────────────────────────────

# Web framework
fastapi==0.133.1
uvicorn==0.41.0
python-multipart==0.0.22
starlette==0.52.1

# HTTP client
httpx==0.28.1

# Environment
python-dotenv==1.2.2

# MongoDB
pymongo==4.16.0
dnspython==2.8.0

# NLP / Embeddings
sentence-transformers==2.6.1
transformers==4.38.2
tokenizers==0.15.2

# PyTorch CPU-only (installed via --extra-index-url in Dockerfile)
torch==2.10.0+cpu
torchvision==0.25.0+cpu

# ML / Science
scikit-learn==1.8.0
scipy==1.17.1
numpy==2.3.5

# Retry logic
tenacity==9.1.4

# Other
pydantic==2.12.5
requests==2.32.5
pillow==12.0.0
huggingface_hub==0.36.2
safetensors==0.7.0
tqdm==4.67.3
regex==2026.2.19
PyYAML==6.0.3
filelock==3.20.0
fsspec==2025.12.0
packaging==26.0
Jinja2==3.1.6
MarkupSafe==3.0.2
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.3
sniffio==1.3.1
anyio==4.12.1
typing_extensions==4.15.0
annotated-types==0.7.0
pydantic_core==2.41.5
typing-inspection==0.4.2
click==8.3.1
h11==0.16.0
httpcore==1.0.9
colorama==0.4.6
joblib==1.5.3
threadpoolctl==3.6.0
mpmath==1.3.0
networkx==3.6.1
sympy==1.14.0
flatbuffers==25.12.19
